Optimizing Step 4:
* Used IntStream.rangeClosed() instead of Stream.iterate()
* Added .parallel() for parallel processing

Matching the checklist:

* New controller without ResponseEntity

Endpoints:

* GET /analytics/students/names-starting-with - names to "A" (default), can be changed
* GET /analytics/students/average-age - average age of students
* GET /analytics/faculties/longest-name - the longest faculty name.
* GET /analytics/compute/original-sum - original sum calculation
* GET /analytics/compute/optimized-sum - optimized sum calculation

Some minor fixes from previous homework at StudentService and FacultyService

+++ src/main/java/ru/hogwarts/school/controller/AnalyticsController.java
@@ -48,7 +48,7 @@ public class AnalyticsController {
     }
 
     @GetMapping("/compute/optimized-sum")
-    public Integer calculateOptimizedSum() {
+    public Long calculateOptimizedSum() {
         logger.info("Was invoked GET endpoint for optimized sum calculation");
         return analyticsService.calculateOptimizedSum();
     }
